requirements.txt
django
django_rest_framework
djangorestframework-simplejwt 
djoser
drf_spectacular